Moves `ClearState` to Listeners namespace
